+++ epoch2/cuda/ee_mumu/src/HelAmps_sm.h
@@ -85,12 +85,6 @@ namespace MG5_sm
 // 
 // 
 
-__device__ void FFV1_0(const cxtype F1[], const cxtype F2[], const cxtype V3[],
-    const cxtype COUP, cxtype * vertex);
-
-__device__ void FFV1P0_3(const cxtype F1[], const cxtype F2[], const cxtype
-    COUP, const fptype M3, const fptype W3, cxtype V3[]);
-
 __device__ void FFV2_0(const cxtype F1[], const cxtype F2[], const cxtype V3[],
     const cxtype COUP, cxtype * vertex);
 
@@ -102,6 +96,12 @@ __device__ void FFV4_0(const cxtype F1[], const cxtype F2[], const cxtype V3[],
 
 __device__ void FFV4_3(const cxtype F1[], const cxtype F2[], const cxtype COUP,
     const fptype M3, const fptype W3, cxtype V3[]);
+
+__device__ void FFV1_0(const cxtype F1[], const cxtype F2[], const cxtype V3[],
+    const cxtype COUP, cxtype * vertex);
+
+__device__ void FFV1P0_3(const cxtype F1[], const cxtype F2[], const cxtype
+    COUP, const fptype M3, const fptype W3, cxtype V3[]);
 
 __device__ void FFV2_4_0(const cxtype F1[], const cxtype F2[], const cxtype
     V3[], const cxtype COUP1, const cxtype COUP2, cxtype * vertex);
