+++ epoch2/cuda/ee_mumu/src/Parameters_sm.cc
@@ -43,9 +43,9 @@ void Parameters_sm::setIndependentParameters(SLHAReader& slha)
   mdl_MTA = slha.get_block_entry("mass", 15, 1.777000e+00); 
   mdl_MT = slha.get_block_entry("mass", 6, 1.730000e+02); 
   mdl_MB = slha.get_block_entry("mass", 5, 4.700000e+00); 
-  mdl_conjg__CKM1x1 = 1.; 
-  mdl_conjg__CKM3x3 = 1.; 
   mdl_CKM3x3 = 1.; 
+  mdl_conjg__CKM3x3 = 1.; 
+  mdl_conjg__CKM1x1 = 1.; 
   mdl_complexi = std::complex<double> (0., 1.); 
   mdl_MZ__exp__2 = ((mdl_MZ) * (mdl_MZ)); 
   mdl_MZ__exp__4 = ((mdl_MZ) * (mdl_MZ) * (mdl_MZ) * (mdl_MZ)); 
@@ -129,12 +129,12 @@ void Parameters_sm::printIndependentParameters()
       setw(10) << mdl_MT << endl;
   cout << setw(20) <<  "mdl_MB " <<  "= " << setiosflags(ios::scientific) <<
       setw(10) << mdl_MB << endl;
-  cout << setw(20) <<  "mdl_conjg__CKM1x1 " <<  "= " <<
-      setiosflags(ios::scientific) << setw(10) << mdl_conjg__CKM1x1 << endl;
-  cout << setw(20) <<  "mdl_conjg__CKM3x3 " <<  "= " <<
-      setiosflags(ios::scientific) << setw(10) << mdl_conjg__CKM3x3 << endl;
   cout << setw(20) <<  "mdl_CKM3x3 " <<  "= " << setiosflags(ios::scientific)
       << setw(10) << mdl_CKM3x3 << endl;
+  cout << setw(20) <<  "mdl_conjg__CKM3x3 " <<  "= " <<
+      setiosflags(ios::scientific) << setw(10) << mdl_conjg__CKM3x3 << endl;
+  cout << setw(20) <<  "mdl_conjg__CKM1x1 " <<  "= " <<
+      setiosflags(ios::scientific) << setw(10) << mdl_conjg__CKM1x1 << endl;
   cout << setw(20) <<  "mdl_complexi " <<  "= " << setiosflags(ios::scientific)
       << setw(10) << mdl_complexi << endl;
   cout << setw(20) <<  "mdl_MZ__exp__2 " <<  "= " <<
