+++ epoch2/cuda/gg_ttgg/SubProcesses/P1_Sigma_sm_gg_ttxgg/CPPProcess.cc
@@ -1,0 +1,1 @@
+gCPPProcess.cu
