--- epoch2/cuda/gg_ttgg/SubProcesses/P1_Sigma_sm_gg_ttxgg/CPPProcess.h
@@ -1,276 +1,0 @@
-//==========================================================================
-// This file has been automatically generated for C++ Standalone by
-// MadGraph5_aMC@NLO v. 2.8.2, 2020-10-30
-// By the MadGraph5_aMC@NLO Development Team
-// Visit launchpad.net/madgraph5 and amcatnlo.web.cern.ch
-//==========================================================================
-
-//==========================================================================
-// This file has been automatically generated for C++ Standalone
-// MadGraph5_aMC@NLO v. 2.8.2, 2020-10-30
-// By the MadGraph5_aMC@NLO Development Team
-// Visit launchpad.net/madgraph5 and amcatnlo.web.cern.ch
-// GPU Template
-//==========================================================================
-
-#ifndef HelAmps_sm_H
-#define HelAmps_sm_H
-
-#include <cmath> 
-#include "mgOnGpuConfig.h"
-#include "mgOnGpuTypes.h"
-
-using namespace std; 
-
-namespace MG5_sm 
-{
-// #ifdef __CUDACC__
-// 
-// __device__ void oxxxxx(const fptype p[3], const fptype& fmass, const int&
-// nhel, const int& nsf,
-// cxtype fo[6]);
-// __device__ void omzxxx(const fptype p[3],  const int& nhel, const int& nsf,
-// cxtype fo[6]);
-// __device__ void opzxxx(const fptype p[3], const int& nhel, const int& nsf,
-// cxtype fo[6]);
-// __device__ void oxzxxx(const fptype p[3],  const int& nhel, const int& nsf,
-// cxtype fo[6]);
-// 
-// __device__ void sxxxxx(const fptype p[3], const int& nss, cxtype sc[3]);
-// 
-// __device__ void ixxxxx(const fptype p[3], const fptype& fmass, const int&
-// nhel, const int& nsf,
-// cxtype fi[6]);
-// __device__ void imzxxx(const fptype p[3], const int& nhel, const int& nsf,
-// cxtype fi[6]);
-// __device__ void ipzxxx(const fptype p[3], const int& nhel, const int& nsf,
-// cxtype fi[6]);
-// __device__ void ixzxxx(const fptype p[3], const int& nhel, const int& nsf,
-// cxtype fi[6]);
-// 
-// __device__ void txxxxx(fptype p[4], fptype tmass, int nhel, int nst,
-// cxtype fi[18]);
-// 
-// __device__ void vxxxxx(const fptype p[3], const fptype& vmass, const int&
-// nhel, const int& nsv,
-// cxtype v[6]);
-// 
-// #else
-// 
-// __device__ void oxxxxx(const fptype p[3], const fptype& fmass, const int&
-// nhel, const int& nsf, const int& ievt,
-// cxtype fo[6]);
-// __device__ void omzxxx(const fptype p[3],  const int& nhel, const int& nsf,
-// cxtype fo[6]);
-// __device__ void opzxxx(const fptype p[3], const int& nhel, const int& nsf,
-// cxtype fo[6]);
-// __device__ void oxzxxx(const fptype p[3],  const int& nhel, const int& nsf,
-// cxtype fo[6]);
-// 
-// __device__ void sxxxxx(const fptype p[3], const int& nss, cxtype sc[3]);
-// 
-// __device__ void ixxxxx(const fptype p[3], const fptype& fmass, const int&
-// nhel, const int& nsf,
-// cxtype fi[6]);
-// __device__ void imzxxx(const fptype p[3], const int& nhel, const int& nsf,
-// cxtype fi[6]);
-// __device__ void ipzxxx(const fptype p[3], const int& nhel, const int& nsf,
-// cxtype fi[6]);
-// __device__ void ixzxxx(const fptype p[3], const int& nhel, const int& nsf,
-// cxtype fi[6]);
-// 
-// __device__ void txxxxx(fptype p[4], fptype tmass, int nhel, int nst,
-// cxtype fi[18]);
-// 
-// __device__ void vxxxxx(const fptype p[3], const fptype& vmass, const int&
-// nhel, const int& nsv,
-// cxtype v[6]);
-// 
-// 
-// 
-// #endif
-// 
-// 
-
-__device__ void FFV1_0(const cxtype F1[], const cxtype F2[], const cxtype V3[], 
-const cxtype COUP, cxtype * vertex); 
-
-__device__ void FFV1_1(const cxtype F2[], const cxtype V3[], const cxtype COUP, 
-const fptype M1, const fptype W1, cxtype F1[]); 
-
-__device__ void FFV1_2(const cxtype F1[], const cxtype V3[], const cxtype COUP, 
-const fptype M2, const fptype W2, cxtype F2[]); 
-
-__device__ void FFV1P0_3(const cxtype F1[], const cxtype F2[], const cxtype
-COUP, const fptype M3, const fptype W3, cxtype V3[]); 
-
-__device__ void VVVV4_0(const cxtype V1[], const cxtype V2[], const cxtype
-V3[], const cxtype V4[], const cxtype COUP, cxtype * vertex); 
-
-__device__ void VVVV4P0_1(const cxtype V2[], const cxtype V3[], const cxtype
-V4[], const cxtype COUP, const fptype M1, const fptype W1, cxtype V1[]); 
-
-__device__ void VVVV3_0(const cxtype V1[], const cxtype V2[], const cxtype
-V3[], const cxtype V4[], const cxtype COUP, cxtype * vertex); 
-
-__device__ void VVVV3P0_1(const cxtype V2[], const cxtype V3[], const cxtype
-V4[], const cxtype COUP, const fptype M1, const fptype W1, cxtype V1[]); 
-
-__device__ void VVV1_0(const cxtype V1[], const cxtype V2[], const cxtype V3[], 
-const cxtype COUP, cxtype * vertex); 
-
-__device__ void VVV1P0_1(const cxtype V2[], const cxtype V3[], const cxtype
-COUP, const fptype M1, const fptype W1, cxtype V1[]); 
-
-__device__ void VVVV1_0(const cxtype V1[], const cxtype V2[], const cxtype
-V3[], const cxtype V4[], const cxtype COUP, cxtype * vertex); 
-
-__device__ void VVVV1P0_1(const cxtype V2[], const cxtype V3[], const cxtype
-V4[], const cxtype COUP, const fptype M1, const fptype W1, cxtype V1[]); 
-
-}  // end namespace MG5_sm
-
-#endif // HelAmps_sm_H
-
-
-#ifndef MG5_Sigma_sm_gg_ttxgg_H
-#define MG5_Sigma_sm_gg_ttxgg_H
-
-#include <complex>
-#include <vector>
-#include <cassert>
-#include <iostream>
-
-#include "mgOnGpuConfig.h"
-#include "mgOnGpuTypes.h"
-
-
-#include "Parameters_sm.h"
-
-//--------------------------------------------------------------------------
-
-#ifdef __CUDACC__
-
-#define checkCuda(code) {assertCuda(code, __FILE__, __LINE__);}
-
-inline void assertCuda(cudaError_t code, const char * file, int line, bool
-    abort = true)
-{
-  if (code != cudaSuccess)
-  {
-    printf( "GPUassert: %s %s:%d\n", cudaGetErrorString(code), file, line); 
-    if (abort)
-      assert(code == cudaSuccess); 
-  }
-}
-
-#endif
-//--------------------------------------------------------------------------
-
-#ifdef __CUDACC__
-namespace gProc
-#else
-namespace Proc
-#endif
-{
-
-//==========================================================================
-// A class for calculating the matrix elements for
-// Process: g g > t t~ g g WEIGHTED<=4 @1
-//--------------------------------------------------------------------------
-
-class CPPProcess
-{
-  public:
-
-    CPPProcess(int numiterations, int gpublocks, int gputhreads, 
-    bool verbose = false, bool debug = false); 
-
-    ~CPPProcess(); 
-
-    // Initialize process.
-    virtual void initProc(std::string param_card_name); 
-
-
-    virtual int code() const {return 1;}
-
-    const std::vector<fptype> &getMasses() const; 
-
-    void setInitial(int inid1, int inid2) 
-    {
-      id1 = inid1; 
-      id2 = inid2; 
-    }
-
-    int getDim() const {return dim;}
-
-    int getNIOParticles() const {return nexternal;}
-
-
-    // Constants for array limits
-    static const int ninitial = mgOnGpu::npari; 
-    static const int nexternal = mgOnGpu::npar; 
-    // static const int nprocesses = 1;
-
-  private:
-    int m_numiterations; 
-    // gpu variables
-    int gpu_nblocks; 
-    int gpu_nthreads; 
-    int dim;  // gpu_nblocks * gpu_nthreads;
-
-    // print verbose info
-    bool m_verbose; 
-
-    // print debug info
-    bool m_debug; 
-
-    static const int nwavefuncs = 6; 
-    static const int namplitudes = 159; 
-    static const int ncomb = 64; 
-    static const int wrows = 63; 
-    // static const int nioparticles = 6;
-
-    cxtype** amp; 
-
-
-    // Pointer to the model parameters
-    Parameters_sm * pars; 
-
-    // vector with external particle masses
-    std::vector<fptype> mME; 
-
-    // Initial particle ids
-    int id1, id2; 
-
-}; 
-
-
-
-//--------------------------------------------------------------------------
-#ifdef __CUDACC__
-__global__
-void sigmaKin_getGoodHel(const fptype * allmomenta,  // input: momenta as AOSOA[npagM][npar][4][neppM] with nevt=npagM*neppM
-bool * isGoodHel);  // output: isGoodHel[ncomb] - device array
-#endif
-
-//--------------------------------------------------------------------------
-
-#ifdef __CUDACC__
-void sigmaKin_setGoodHel(const bool * isGoodHel);  // input: isGoodHel[ncomb] - host array
-#endif
-
-//--------------------------------------------------------------------------
-
-__global__
-void sigmaKin(const fptype * allmomenta,  // input: momenta as AOSOA[npagM][npar][4][neppM] with nevt=npagM*neppM
-fptype * allMEs  // output: allMEs[nevt], final |M|^2 averaged over all helicities
-#ifndef __CUDACC__
-, const int nevt  // input: #events (for cuda: nevt == ndim == gpublocks*gputhreads)
-#endif
-); 
-
-//--------------------------------------------------------------------------
-}
-
-#endif // MG5_Sigma_sm_gg_ttxgg_H
