+++ epoch2/cuda/gg_ttgg/SubProcesses/P1_Sigma_sm_gg_ttxgg/check_sa.cc
@@ -1,0 +1,1 @@
+gcheck_sa.cu
