+++ epoch2/cuda/gg_ttgg/src/HelAmps_sm.h
@@ -91,11 +91,11 @@ __device__ void VVVV1_0(const cxtype V1[], const cxtype V2[], const cxtype
 __device__ void VVVV1P0_1(const cxtype V2[], const cxtype V3[], const cxtype
     V4[], const cxtype COUP, const fptype M1, const fptype W1, cxtype V1[]);
 
-__device__ void VVV1_0(const cxtype V1[], const cxtype V2[], const cxtype V3[],
-    const cxtype COUP, cxtype * vertex);
+__device__ void VVVV4_0(const cxtype V1[], const cxtype V2[], const cxtype
+    V3[], const cxtype V4[], const cxtype COUP, cxtype * vertex);
 
-__device__ void VVV1P0_1(const cxtype V2[], const cxtype V3[], const cxtype
-    COUP, const fptype M1, const fptype W1, cxtype V1[]);
+__device__ void VVVV4P0_1(const cxtype V2[], const cxtype V3[], const cxtype
+    V4[], const cxtype COUP, const fptype M1, const fptype W1, cxtype V1[]);
 
 __device__ void VVVV3_0(const cxtype V1[], const cxtype V2[], const cxtype
     V3[], const cxtype V4[], const cxtype COUP, cxtype * vertex);
@@ -115,11 +115,11 @@ __device__ void FFV1_2(const cxtype F1[], const cxtype V3[], const cxtype COUP,
 __device__ void FFV1P0_3(const cxtype F1[], const cxtype F2[], const cxtype
     COUP, const fptype M3, const fptype W3, cxtype V3[]);
 
-__device__ void VVVV4_0(const cxtype V1[], const cxtype V2[], const cxtype
-    V3[], const cxtype V4[], const cxtype COUP, cxtype * vertex);
+__device__ void VVV1_0(const cxtype V1[], const cxtype V2[], const cxtype V3[],
+    const cxtype COUP, cxtype * vertex);
 
-__device__ void VVVV4P0_1(const cxtype V2[], const cxtype V3[], const cxtype
-    V4[], const cxtype COUP, const fptype M1, const fptype W1, cxtype V1[]);
+__device__ void VVV1P0_1(const cxtype V2[], const cxtype V3[], const cxtype
+    COUP, const fptype M1, const fptype W1, cxtype V1[]);
 
 }  // end namespace MG5_sm
